+++ shared-components/src/components/button/Button.tsx
@@ -1,17 +1,23 @@
-import { FC } from "react";
+import { forwardRef } from "react";
 import { ButtonProps, BaseButton } from "./BaseButton";
 
 import "./button.scss";
 import "./button-shared.scss";
 
-export const PrimaryButton: FC<ButtonProps> = (props) => (
-  <BaseButton buttonStyle="rds-button" buttonType="primary" {...props} />
-);
+export const PrimaryButton = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) => (
+  <BaseButton buttonStyle="rds-button" buttonType="primary" ref={ref} {...props} />
+));
 
-export const SecondaryButton: FC<ButtonProps> = (props) => (
-  <BaseButton buttonStyle="rds-button" buttonType="secondary" {...props} />
-);
+PrimaryButton.displayName = "PrimaryButton";
 
-export const TertiaryButton: FC<ButtonProps> = (props) => (
-  <BaseButton buttonStyle="rds-button" buttonType="tertiary" {...props} />
-);
+export const SecondaryButton = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) => (
+  <BaseButton buttonStyle="rds-button" buttonType="secondary" ref={ref} {...props} />
+));
+
+SecondaryButton.displayName = "SecondaryButton";
+
+export const TertiaryButton = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) => (
+  <BaseButton buttonStyle="rds-button" buttonType="tertiary" ref={ref} {...props} />
+));
+
+TertiaryButton.displayName = "TertiaryButton";